Simple Galaxy Gate: Move GateBuilder, KamikazeHandler and StateStore to utils package
