+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/SimpleGalaxyGate.java
@@ -14,7 +14,9 @@ import com.github.manolo8.darkbot.core.itf.NpcExtraProvider;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.GateNpcFlag;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.Maps;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.SimpleGalaxyGateConfig;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateBuilder;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import dev.shared.do_gamer.utils.BackpageHelper;
 import dev.shared.do_gamer.utils.PetGearHelper;
 import dev.shared.do_gamer.utils.ServerTimeHelper;
