+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/gate/DseGate.java
@@ -1,7 +1,7 @@
 package dev.shared.do_gamer.module.simple_galaxy_gate.gate;
 
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import eu.darkbot.api.config.types.NpcFlag;
 import eu.darkbot.api.game.entities.Npc;
 import eu.darkbot.api.game.other.GameMap;
