+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/gate/GopGate.java
@@ -2,8 +2,8 @@ package dev.shared.do_gamer.module.simple_galaxy_gate.gate;
 
 import java.util.Comparator;
 
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import eu.darkbot.api.PluginAPI;
 import eu.darkbot.api.config.types.NpcFlag;
 import eu.darkbot.api.game.entities.Npc;
