+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/gate/InvasionGate.java
@@ -2,8 +2,8 @@ package dev.shared.do_gamer.module.simple_galaxy_gate.gate;
 
 import java.util.List;
 
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import eu.darkbot.api.game.enums.PortalType;
 import eu.darkbot.api.game.other.GameMap;
