+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/gate/TreacherousDomainGate.java
@@ -3,9 +3,9 @@ package dev.shared.do_gamer.module.simple_galaxy_gate.gate;
 import java.time.LocalDateTime;
 import java.util.Comparator;
 
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.ScheduledGateHelper;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import dev.shared.do_gamer.utils.ServerTimeHelper;
 import eu.darkbot.api.game.entities.Npc;
