+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/gate/ZetaGate.java
@@ -2,8 +2,8 @@ package dev.shared.do_gamer.module.simple_galaxy_gate.gate;
 
 import com.github.manolo8.darkbot.core.entities.Npc;
 
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
+import dev.shared.do_gamer.module.simple_galaxy_gate.utils.StateStore;
 import eu.darkbot.api.config.types.NpcFlag;
 
 public final class ZetaGate extends GateHandler {
