+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/utils/KamikazeHandler.java
@@ -1,4 +1,4 @@
-package dev.shared.do_gamer.module.simple_galaxy_gate;
+package dev.shared.do_gamer.module.simple_galaxy_gate.utils;
 
 import java.util.Comparator;
 import java.util.List;
@@ -6,10 +6,10 @@ import java.util.Map;
 import java.util.WeakHashMap;
 import java.util.stream.Collectors;
 
+import dev.shared.do_gamer.module.simple_galaxy_gate.CustomLootModule;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.GateNpcFlag;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.Maps;
 import dev.shared.do_gamer.module.simple_galaxy_gate.config.SimpleGalaxyGateConfig;
-import dev.shared.do_gamer.module.simple_galaxy_gate.utils.GateHandler;
 import dev.shared.do_gamer.utils.PetGearHelper;
 import eu.darkbot.api.PluginAPI;
 import eu.darkbot.api.game.entities.Npc;
