+++ src/main/java/dev/shared/do_gamer/module/simple_galaxy_gate/utils/ScheduledGateHelper.java
@@ -6,7 +6,6 @@ import java.util.function.LongSupplier;
 import java.util.function.LongUnaryOperator;
 
 import dev.shared.do_gamer.module.simple_galaxy_gate.SimpleGalaxyGate;
-import dev.shared.do_gamer.module.simple_galaxy_gate.StateStore;
 import eu.darkbot.util.Timer;
 
 /**
